requirements.txt
fastapi==0.115.5
uvicorn
pymongo==4.10.1
python-dotenv
langchain_groq
langchain
langchain-community
beautifulsoup4==4.12.3
pytest
pytest-asyncio
httpx
pinecone[grpc]
